requirements.txt
ultralytics==8.0.196
streamlit
opencv-python
numpy
Pillow
matplotlib
pandas
